requirements.txt
Django==1.9.4
django-ckeditor==5.0.3
django-mptt==0.8.4
Pillow==3.1.1
pytz==2016.3
